Move all contextMenu methods out of localModel, into ChartModel.ts
